fix: fix data type in ServiceImpl and Request

+++ psgp_be/src/main/java/com/exe201/group1/psgp_be/repositories/SupplierRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SupplierRepo extends JpaRepository<Supplier, Long> {
+public interface SupplierRepo extends JpaRepository<Supplier, Integer> {
     boolean existsBySupplierNameIgnoreCase(String supplierName);
     boolean existsBySupplierNameIgnoreCaseAndIdNot(String supplierName, int id);
 }
